requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-ckeditor==6.2.0
django-import-export==2.7.0
django-js-asset==1.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.9.2
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
xlrd==2.0.1
xlwt==1.3.0
